This PR improves the filter kernel:

* made the filter benchmarks more realistic
* performance improved by 1.2-200x
* filter now supports all types supported by `MutableArrayData`
* removed 500 LOC

There are two novel ideas here:

1. it minimizes the number of re-allocations when building the filtered array, both for single filter and multi-filter operations.

2. For single filter operations, it leverages iterators to create the new array on the fly. For multi filter operations, it persists the iterator's result in a vector and iterates over it per array.

This PR also improves the performance of `MutableArrayData` by avoiding some bound checks via `unsafe` (properly documented).

Summary of the benchmarks:

|  benchmark | variation (%) |
|-------------- | -------------- | 
| filter context string high selectivity | -26.3 | 
| filter context u8 w NULLs high selectivity | -26.8 | 
| filter context f32 high selectivity | -28.0 | 
| filter u8 low selectivity | -43.5 | 
| filter u8 high selectivity | -66.5 | 
| filter context u8 w NULLs low selectivity | -72.7 | 
| filter context u8 low selectivity | -73.5 | 
| filter context f32 low selectivity | -77.7 | 
| filter context string low selectivity | -78.6 | 
| filter context u8 high selectivity | -78.8 | 
| filter u8 | -98.7 | 
| filter context f32 | -99.3 | 
| filter context string | -99.4 | 
| filter context u8 | -99.4 | 
| filter context u8 w NULLs | -99.5 |

Code used to benchmark:
```bash
git checkout 54da4378d138df12bd4e09a68b0f4c80218834c7
cargo bench --bench filter_kernels
git checkout mutable_filter2
cargo bench --bench filter_kernels
```

Benchmark result:
```
Switched to branch 'mutable_filter2'
   Compiling arrow v3.0.0-SNAPSHOT (/Users/jorgecarleitao/projects/arrow/rust/arrow)
    Finished bench [optimized] target(s) in 1m 01s
     Running /Users/jorgecarleitao/projects/arrow/rust/target/release/deps/filter_kernels-5208f9a404de52c9
Gnuplot not found, using plotters backend
filter u8               time:   [5.2289 us 5.4746 us 5.7646 us]                       
                        change: [-98.691% -98.666% -98.637%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  3 (3.00%) high mild
  13 (13.00%) high severe

filter u8 high selectivity                                                                             
                        time:   [7.2731 us 7.2838 us 7.2953 us]
                        change: [-66.869% -66.505% -66.140%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe

filter u8 low selectivity                                                                             
                        time:   [3.6938 us 3.6993 us 3.7046 us]
                        change: [-43.987% -43.525% -43.104%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

filter context u8       time:   [2.3932 us 2.3973 us 2.4013 us]                               
                        change: [-99.392% -99.388% -99.384%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  3 (3.00%) low mild
  2 (2.00%) high mild
  7 (7.00%) high severe

filter context u8 high selectivity                                                                             
                        time:   [4.2641 us 4.2726 us 4.2828 us]
                        change: [-79.030% -78.771% -78.497%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) high mild
  8 (8.00%) high severe

filter context u8 low selectivity                                                                             
                        time:   [1.3684 us 1.3735 us 1.3787 us]
                        change: [-73.826% -73.495% -73.208%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

filter context u8 w NULLs                                                                             
                        time:   [2.7073 us 2.7111 us 2.7151 us]
                        change: [-99.540% -99.535% -99.530%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  9 (9.00%) high severe

filter context u8 w NULLs high selectivity                                                                            
                        time:   [280.66 us 290.65 us 306.21 us]
                        change: [-28.378% -26.780% -24.548%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  7 (7.00%) high mild
  9 (9.00%) high severe

filter context u8 w NULLs low selectivity                                                                             
                        time:   [1.6484 us 1.6536 us 1.6603 us]
                        change: [-72.983% -72.674% -72.356%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  11 (11.00%) high severe

filter context f32      time:   [4.6148 us 4.6205 us 4.6259 us]                                 
                        change: [-99.262% -99.251% -99.237%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  1 (1.00%) low severe
  4 (4.00%) low mild
  1 (1.00%) high mild
  13 (13.00%) high severe

filter context f32 high selectivity                                                                            
                        time:   [285.17 us 286.05 us 287.16 us]
                        change: [-28.812% -28.042% -27.358%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  3 (3.00%) high mild
  10 (10.00%) high severe

filter context f32 low selectivity                                                                             
                        time:   [1.5320 us 1.5450 us 1.5648 us]
                        change: [-77.882% -77.690% -77.468%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

filter context string   time:   [6.0464 us 6.0541 us 6.0614 us]                                   
                        change: [-99.362% -99.357% -99.350%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high severe

Benchmarking filter context string high selectivity: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.2s, enable flat sampling, or reduce sample count to 60.
filter context string high selectivity                                                                             
                        time:   [936.97 us 940.91 us 945.24 us]
                        change: [-26.991% -26.325% -25.679%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  10 (10.00%) high mild
  6 (6.00%) high severe

filter context string low selectivity                                                                             
                        time:   [1.7669 us 1.7707 us 1.7749 us]
                        change: [-79.102% -78.624% -78.201%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
```