@mqy 2020-12-17
Amazing, great work! Here is my bench result on mac book pro 2018:

$ git checkout mutable_filter2
Previous HEAD position was 54da4378d Improved filter bench.
Switched to branch 'mutable_filter2'
Your branch is up to date with 'jorgecarleitao/mutable_filter2'.
$ cargo bench --bench filter_kernels
   Compiling arrow v3.0.0-SNAPSHOT (/Users/mqy/tools/data/arrow/rust/arrow)
    Finished bench [optimized] target(s) in 30.81s
     Running target/release/deps/filter_kernels-9a777356dd1c1b39
Gnuplot not found, using plotters backend
filter u8               time:   [3.3223 us 3.3282 us 3.3340 us]                       
                        change: [**-99.062% -99.059% -99.056%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  3 (3.00%) high severe

filter u8 high selectivity                                                                             
                        time:   [6.6873 us 6.8811 us 7.0893 us]
                        change: [**-65.458% -64.142% -62.812%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  7 (7.00%) high mild

filter u8 low selectivity                                                                             
                        time:   [2.6626 us 2.6704 us 2.6783 us]
                        change: [**-53.949% -53.755% -53.558%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

filter context u8       time:   [1.6725 us 1.6814 us 1.6902 us]                               
                        change: [**-99.525% -99.522% -99.519%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild

filter context u8 high selectivity                                                                             
                        time:   [4.3977 us 4.5770 us 4.7594 us]
                        change: [**-75.323% -73.815% -72.286%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  11 (11.00%) high mild

filter context u8 low selectivity                                                                             
                        time:   [1.1797 us 1.1843 us 1.1890 us]
                        change: [**-74.379% -74.076% -73.824%**] (p = 0.00 < 0.05)
                        Performance has improved.

filter context u8 w NULLs                                                                             
                        time:   [2.0055 us 2.0120 us 2.0187 us]
                        change: [**-99.621% -99.619% -99.618%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild

filter context u8 w NULLs high selectivity                                                                            
                        time:   [243.76 us 244.08 us 244.41 us]
                        change: [**-32.511% -32.318% -32.144%**] (p = 0.00 < 0.05)
                        Performance has improved.

filter context u8 w NULLs low selectivity                                                                             
                        time:   [1.4073 us 1.4151 us 1.4230 us]
                        change: [**-73.174% -72.860% -72.574%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

filter context f32      time:   [3.7693 us 3.7804 us 3.7918 us]                                
                        change: [**-99.341% -99.337% -99.332%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

filter context f32 high selectivity                                                                            
                        time:   [257.52 us 258.42 us 259.27 us]
                        change: [**-31.098% -30.752% -30.429%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

filter context f32 low selectivity                                                                             
                        time:   [1.3838 us 1.3901 us 1.3964 us]
                        change: [**-77.474% -77.227% -77.021%**] (p = 0.00 < 0.05)
                        Performance has improved.

filter context string   time:   [4.7862 us 4.8040 us 4.8206 us]                                   
                        change: [**-99.393% -99.389% -99.386%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
  3 (3.00%) high severe

filter context string high selectivity                                                                            
                        time:   [833.56 us 836.78 us 843.25 us]
                        change: [**-20.015% -19.639% -19.202%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  4 (4.00%) high severe

filter context string low selectivity                                                                             
                        time:   [1.6389 us 1.6486 us 1.6584 us]
                        change: [**-76.728% -76.552% -76.376%**] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild